split test_basic more
