+++ python/pyarrow/tests/parquet/test_basic.py
@@ -32,7 +32,7 @@ from pyarrow.tests import util
 from pyarrow.tests.parquet.common import (
     _check_roundtrip, _roundtrip_table, make_sample_file,
     parametrize_legacy_dataset, parametrize_legacy_dataset_fixed,
-    parametrize_legacy_dataset_not_supported)
+    parametrize_legacy_dataset_not_supported, _simple_table_write_read)
 
 try:
     import pyarrow.parquet as pq
@@ -105,28 +105,6 @@ def test_single_pylist_column_roundtrip(tempdir, dtype, use_legacy_dataset):
         data_read = col_read.chunk(0)
         assert data_written.equals(data_read)
 
-
-@pytest.mark.pandas
-@parametrize_legacy_dataset
-@pytest.mark.parametrize('chunk_size', [None, 1000])
-def test_pandas_parquet_2_0_roundtrip(tempdir, chunk_size, use_legacy_dataset):
-    df = alltypes_sample(size=10000, categorical=True)
-
-    filename = tempdir / 'pandas_roundtrip.parquet'
-    arrow_table = pa.Table.from_pandas(df)
-    assert arrow_table.schema.pandas_metadata is not None
-
-    _write_table(arrow_table, filename, version="2.0",
-                 coerce_timestamps='ms', chunk_size=chunk_size)
-    table_read = pq.read_pandas(
-        filename, use_legacy_dataset=use_legacy_dataset)
-    assert table_read.schema.pandas_metadata is not None
-
-    read_metadata = table_read.schema.metadata
-    assert arrow_table.schema.metadata == read_metadata
-
-    df_read = table_read.to_pandas()
-    tm.assert_frame_equal(df, df_read)
 
 
 def test_parquet_invalid_version(tempdir):
@@ -277,146 +255,6 @@ def test_nested_list_nonnullable_roundtrip_bug(use_legacy_dataset):
 
 @pytest.mark.pandas
 @parametrize_legacy_dataset
-def test_pandas_parquet_datetime_tz(use_legacy_dataset):
-    s = pd.Series([datetime.datetime(2017, 9, 6)])
-    s = s.dt.tz_localize('utc')
-
-    s.index = s
-
-    # Both a column and an index to hit both use cases
-    df = pd.DataFrame({'tz_aware': s,
-                       'tz_eastern': s.dt.tz_convert('US/Eastern')},
-                      index=s)
-
-    f = io.BytesIO()
-
-    arrow_table = pa.Table.from_pandas(df)
-
-    _write_table(arrow_table, f, coerce_timestamps='ms')
-    f.seek(0)
-
-    table_read = pq.read_pandas(f, use_legacy_dataset=use_legacy_dataset)
-
-    df_read = table_read.to_pandas()
-    tm.assert_frame_equal(df, df_read)
-
-
-@pytest.mark.pandas
-@parametrize_legacy_dataset
-def test_datetime_timezone_tzinfo(use_legacy_dataset):
-    value = datetime.datetime(2018, 1, 1, 1, 23, 45,
-                              tzinfo=datetime.timezone.utc)
-    df = pd.DataFrame({'foo': [value]})
-
-    _roundtrip_pandas_dataframe(
-        df, write_kwargs={}, use_legacy_dataset=use_legacy_dataset)
-
-
-@pytest.mark.pandas
-def test_pandas_parquet_custom_metadata(tempdir):
-    df = alltypes_sample(size=10000)
-
-    filename = tempdir / 'pandas_roundtrip.parquet'
-    arrow_table = pa.Table.from_pandas(df)
-    assert b'pandas' in arrow_table.schema.metadata
-
-    _write_table(arrow_table, filename, version='2.0', coerce_timestamps='ms')
-
-    metadata = pq.read_metadata(filename).metadata
-    assert b'pandas' in metadata
-
-    js = json.loads(metadata[b'pandas'].decode('utf8'))
-    assert js['index_columns'] == [{'kind': 'range',
-                                    'name': None,
-                                    'start': 0, 'stop': 10000,
-                                    'step': 1}]
-
-
-@pytest.mark.pandas
-@parametrize_legacy_dataset
-def test_pandas_parquet_column_multiindex(tempdir, use_legacy_dataset):
-    df = alltypes_sample(size=10)
-    df.columns = pd.MultiIndex.from_tuples(
-        list(zip(df.columns, df.columns[::-1])),
-        names=['level_1', 'level_2']
-    )
-
-    filename = tempdir / 'pandas_roundtrip.parquet'
-    arrow_table = pa.Table.from_pandas(df)
-    assert arrow_table.schema.pandas_metadata is not None
-
-    _write_table(arrow_table, filename, version='2.0', coerce_timestamps='ms')
-
-    table_read = pq.read_pandas(
-        filename, use_legacy_dataset=use_legacy_dataset)
-    df_read = table_read.to_pandas()
-    tm.assert_frame_equal(df, df_read)
-
-
-@pytest.mark.pandas
-@parametrize_legacy_dataset
-def test_pandas_parquet_2_0_roundtrip_read_pandas_no_index_written(
-    tempdir, use_legacy_dataset
-):
-    df = alltypes_sample(size=10000)
-
-    filename = tempdir / 'pandas_roundtrip.parquet'
-    arrow_table = pa.Table.from_pandas(df, preserve_index=False)
-    js = arrow_table.schema.pandas_metadata
-    assert not js['index_columns']
-    # ARROW-2170
-    # While index_columns should be empty, columns needs to be filled still.
-    assert js['columns']
-
-    _write_table(arrow_table, filename, version='2.0', coerce_timestamps='ms')
-    table_read = pq.read_pandas(
-        filename, use_legacy_dataset=use_legacy_dataset)
-
-    js = table_read.schema.pandas_metadata
-    assert not js['index_columns']
-
-    read_metadata = table_read.schema.metadata
-    assert arrow_table.schema.metadata == read_metadata
-
-    df_read = table_read.to_pandas()
-    tm.assert_frame_equal(df, df_read)
-
-
-@pytest.mark.pandas
-@parametrize_legacy_dataset
-def test_pandas_parquet_1_0_roundtrip(tempdir, use_legacy_dataset):
-    size = 10000
-    np.random.seed(0)
-    df = pd.DataFrame({
-        'uint8': np.arange(size, dtype=np.uint8),
-        'uint16': np.arange(size, dtype=np.uint16),
-        'uint32': np.arange(size, dtype=np.uint32),
-        'uint64': np.arange(size, dtype=np.uint64),
-        'int8': np.arange(size, dtype=np.int16),
-        'int16': np.arange(size, dtype=np.int16),
-        'int32': np.arange(size, dtype=np.int32),
-        'int64': np.arange(size, dtype=np.int64),
-        'float32': np.arange(size, dtype=np.float32),
-        'float64': np.arange(size, dtype=np.float64),
-        'bool': np.random.randn(size) > 0,
-        'str': [str(x) for x in range(size)],
-        'str_with_nulls': [None] + [str(x) for x in range(size - 2)] + [None],
-        'empty_str': [''] * size
-    })
-    filename = tempdir / 'pandas_roundtrip.parquet'
-    arrow_table = pa.Table.from_pandas(df)
-    _write_table(arrow_table, filename, version='1.0')
-    table_read = _read_table(filename, use_legacy_dataset=use_legacy_dataset)
-    df_read = table_read.to_pandas()
-
-    # We pass uint32_t as int64_t if we write Parquet version 1.0
-    df['uint32'] = df['uint32'].values.astype(np.int64)
-
-    tm.assert_frame_equal(df, df_read)
-
-
-@pytest.mark.pandas
-@parametrize_legacy_dataset
 def test_multiple_path_types(tempdir, use_legacy_dataset):
     # Test compatibility with PEP 519 path-like objects
     path = tempdir / 'zzz.parquet'
@@ -465,48 +303,6 @@ def test_read_non_existing_file(use_legacy_dataset):
         pq.read_table('i-am-not-existing.parquet')
 
 
-# TODO(dataset) duplicate column selection actually gives duplicate columns now
-@pytest.mark.pandas
-@parametrize_legacy_dataset_not_supported
-def test_pandas_column_selection(tempdir, use_legacy_dataset):
-    size = 10000
-    np.random.seed(0)
-    df = pd.DataFrame({
-        'uint8': np.arange(size, dtype=np.uint8),
-        'uint16': np.arange(size, dtype=np.uint16)
-    })
-    filename = tempdir / 'pandas_roundtrip.parquet'
-    arrow_table = pa.Table.from_pandas(df)
-    _write_table(arrow_table, filename)
-    table_read = _read_table(
-        filename, columns=['uint8'], use_legacy_dataset=use_legacy_dataset)
-    df_read = table_read.to_pandas()
-
-    tm.assert_frame_equal(df[['uint8']], df_read)
-
-    # ARROW-4267: Selection of duplicate columns still leads to these columns
-    # being read uniquely.
-    table_read = _read_table(
-        filename, columns=['uint8', 'uint8'],
-        use_legacy_dataset=use_legacy_dataset)
-    df_read = table_read.to_pandas()
-
-    tm.assert_frame_equal(df[['uint8']], df_read)
-
-
-@pytest.mark.pandas
-@parametrize_legacy_dataset
-def test_pandas_parquet_native_file_roundtrip(tempdir, use_legacy_dataset):
-    df = _test_dataframe(10000)
-    arrow_table = pa.Table.from_pandas(df)
-    imos = pa.BufferOutputStream()
-    _write_table(arrow_table, imos, version="2.0")
-    buf = imos.getvalue()
-    reader = pa.BufferReader(buf)
-    df_read = _read_table(
-        reader, use_legacy_dataset=use_legacy_dataset).to_pandas()
-    tm.assert_frame_equal(df, df_read)
-
 
 @parametrize_legacy_dataset
 def test_parquet_read_from_buffer(tempdir, use_legacy_dataset):
@@ -551,127 +347,6 @@ def test_parquet_incremental_file_build(tempdir, use_legacy_dataset):
 
     expected = pd.concat(frames, ignore_index=True)
     tm.assert_frame_equal(result.to_pandas(), expected)
-
-
-@pytest.mark.pandas
-@parametrize_legacy_dataset
-def test_read_pandas_column_subset(tempdir, use_legacy_dataset):
-    df = _test_dataframe(10000)
-    arrow_table = pa.Table.from_pandas(df)
-    imos = pa.BufferOutputStream()
-    _write_table(arrow_table, imos, version="2.0")
-    buf = imos.getvalue()
-    reader = pa.BufferReader(buf)
-    df_read = pq.read_pandas(
-        reader, columns=['strings', 'uint8'],
-        use_legacy_dataset=use_legacy_dataset
-    ).to_pandas()
-    tm.assert_frame_equal(df[['strings', 'uint8']], df_read)
-
-
-@pytest.mark.pandas
-@parametrize_legacy_dataset
-def test_pandas_parquet_empty_roundtrip(tempdir, use_legacy_dataset):
-    df = _test_dataframe(0)
-    arrow_table = pa.Table.from_pandas(df)
-    imos = pa.BufferOutputStream()
-    _write_table(arrow_table, imos, version="2.0")
-    buf = imos.getvalue()
-    reader = pa.BufferReader(buf)
-    df_read = _read_table(
-        reader, use_legacy_dataset=use_legacy_dataset).to_pandas()
-    tm.assert_frame_equal(df, df_read)
-
-
-@pytest.mark.pandas
-def test_pandas_can_write_nested_data(tempdir):
-    data = {
-        "agg_col": [
-            {"page_type": 1},
-            {"record_type": 1},
-            {"non_consecutive_home": 0},
-        ],
-        "uid_first": "1001"
-    }
-    df = pd.DataFrame(data=data)
-    arrow_table = pa.Table.from_pandas(df)
-    imos = pa.BufferOutputStream()
-    # This succeeds under V2
-    _write_table(arrow_table, imos)
-
-
-@pytest.mark.pandas
-@parametrize_legacy_dataset
-def test_pandas_parquet_pyfile_roundtrip(tempdir, use_legacy_dataset):
-    filename = tempdir / 'pandas_pyfile_roundtrip.parquet'
-    size = 5
-    df = pd.DataFrame({
-        'int64': np.arange(size, dtype=np.int64),
-        'float32': np.arange(size, dtype=np.float32),
-        'float64': np.arange(size, dtype=np.float64),
-        'bool': np.random.randn(size) > 0,
-        'strings': ['foo', 'bar', None, 'baz', 'qux']
-    })
-
-    arrow_table = pa.Table.from_pandas(df)
-
-    with filename.open('wb') as f:
-        _write_table(arrow_table, f, version="1.0")
-
-    data = io.BytesIO(filename.read_bytes())
-
-    table_read = _read_table(data, use_legacy_dataset=use_legacy_dataset)
-    df_read = table_read.to_pandas()
-    tm.assert_frame_equal(df, df_read)
-
-
-@pytest.mark.pandas
-@parametrize_legacy_dataset
-def test_pandas_parquet_configuration_options(tempdir, use_legacy_dataset):
-    size = 10000
-    np.random.seed(0)
-    df = pd.DataFrame({
-        'uint8': np.arange(size, dtype=np.uint8),
-        'uint16': np.arange(size, dtype=np.uint16),
-        'uint32': np.arange(size, dtype=np.uint32),
-        'uint64': np.arange(size, dtype=np.uint64),
-        'int8': np.arange(size, dtype=np.int16),
-        'int16': np.arange(size, dtype=np.int16),
-        'int32': np.arange(size, dtype=np.int32),
-        'int64': np.arange(size, dtype=np.int64),
-        'float32': np.arange(size, dtype=np.float32),
-        'float64': np.arange(size, dtype=np.float64),
-        'bool': np.random.randn(size) > 0
-    })
-    filename = tempdir / 'pandas_roundtrip.parquet'
-    arrow_table = pa.Table.from_pandas(df)
-
-    for use_dictionary in [True, False]:
-        _write_table(arrow_table, filename, version='2.0',
-                     use_dictionary=use_dictionary)
-        table_read = _read_table(
-            filename, use_legacy_dataset=use_legacy_dataset)
-        df_read = table_read.to_pandas()
-        tm.assert_frame_equal(df, df_read)
-
-    for write_statistics in [True, False]:
-        _write_table(arrow_table, filename, version='2.0',
-                     write_statistics=write_statistics)
-        table_read = _read_table(filename,
-                                 use_legacy_dataset=use_legacy_dataset)
-        df_read = table_read.to_pandas()
-        tm.assert_frame_equal(df, df_read)
-
-    for compression in ['NONE', 'SNAPPY', 'GZIP', 'LZ4', 'ZSTD']:
-        if (compression != 'NONE' and
-                not pa.lib.Codec.is_available(compression)):
-            continue
-        _write_table(arrow_table, filename, version='2.0',
-                     compression=compression)
-        table_read = _read_table(
-            filename, use_legacy_dataset=use_legacy_dataset)
-        df_read = table_read.to_pandas()
-        tm.assert_frame_equal(df, df_read)
 
 
 @parametrize_legacy_dataset
@@ -817,79 +492,6 @@ def test_column_of_arrays(tempdir):
 
 
 @pytest.mark.pandas
-def test_coerce_timestamps(tempdir):
-    from collections import OrderedDict
-
-    # ARROW-622
-    arrays = OrderedDict()
-    fields = [pa.field('datetime64',
-                       pa.list_(pa.timestamp('ms')))]
-    arrays['datetime64'] = [
-        np.array(['2007-07-13T01:23:34.123456789',
-                  None,
-                  '2010-08-13T05:46:57.437699912'],
-                 dtype='datetime64[ms]'),
-        None,
-        None,
-        np.array(['2007-07-13T02',
-                  None,
-                  '2010-08-13T05:46:57.437699912'],
-                 dtype='datetime64[ms]'),
-    ]
-
-    df = pd.DataFrame(arrays)
-    schema = pa.schema(fields)
-
-    filename = tempdir / 'pandas_roundtrip.parquet'
-    arrow_table = pa.Table.from_pandas(df, schema=schema)
-
-    _write_table(arrow_table, filename, version="2.0", coerce_timestamps='us')
-    table_read = _read_table(filename)
-    df_read = table_read.to_pandas()
-
-    df_expected = df.copy()
-    for i, x in enumerate(df_expected['datetime64']):
-        if isinstance(x, np.ndarray):
-            df_expected['datetime64'][i] = x.astype('M8[us]')
-
-    tm.assert_frame_equal(df_expected, df_read)
-
-    with pytest.raises(ValueError):
-        _write_table(arrow_table, filename, version='2.0',
-                     coerce_timestamps='unknown')
-
-
-@pytest.mark.pandas
-def test_coerce_timestamps_truncated(tempdir):
-    """
-    ARROW-2555: Test that we can truncate timestamps when coercing if
-    explicitly allowed.
-    """
-    dt_us = datetime.datetime(year=2017, month=1, day=1, hour=1, minute=1,
-                              second=1, microsecond=1)
-    dt_ms = datetime.datetime(year=2017, month=1, day=1, hour=1, minute=1,
-                              second=1)
-
-    fields_us = [pa.field('datetime64', pa.timestamp('us'))]
-    arrays_us = {'datetime64': [dt_us, dt_ms]}
-
-    df_us = pd.DataFrame(arrays_us)
-    schema_us = pa.schema(fields_us)
-
-    filename = tempdir / 'pandas_truncated.parquet'
-    table_us = pa.Table.from_pandas(df_us, schema=schema_us)
-
-    _write_table(table_us, filename, version="2.0", coerce_timestamps='ms',
-                 allow_truncated_timestamps=True)
-    table_ms = _read_table(filename)
-    df_ms = table_ms.to_pandas()
-
-    arrays_expected = {'datetime64': [dt_ms, dt_ms]}
-    df_expected = pd.DataFrame(arrays_expected)
-    tm.assert_frame_equal(df_expected, df_ms)
-
-
-@pytest.mark.pandas
 def test_column_of_lists(tempdir):
     df, schema = dataframe_with_lists(parquet_compatible=True)
 
@@ -901,188 +503,6 @@ def test_column_of_lists(tempdir):
 
     tm.assert_frame_equal(df, df_read)
 
-
-@pytest.mark.pandas
-def test_date_time_types(tempdir):
-    t1 = pa.date32()
-    data1 = np.array([17259, 17260, 17261], dtype='int32')
-    a1 = pa.array(data1, type=t1)
-
-    t2 = pa.date64()
-    data2 = data1.astype('int64') * 86400000
-    a2 = pa.array(data2, type=t2)
-
-    t3 = pa.timestamp('us')
-    start = pd.Timestamp('2001-01-01').value / 1000
-    data3 = np.array([start, start + 1, start + 2], dtype='int64')
-    a3 = pa.array(data3, type=t3)
-
-    t4 = pa.time32('ms')
-    data4 = np.arange(3, dtype='i4')
-    a4 = pa.array(data4, type=t4)
-
-    t5 = pa.time64('us')
-    a5 = pa.array(data4.astype('int64'), type=t5)
-
-    t6 = pa.time32('s')
-    a6 = pa.array(data4, type=t6)
-
-    ex_t6 = pa.time32('ms')
-    ex_a6 = pa.array(data4 * 1000, type=ex_t6)
-
-    t7 = pa.timestamp('ns')
-    start = pd.Timestamp('2001-01-01').value
-    data7 = np.array([start, start + 1000, start + 2000],
-                     dtype='int64')
-    a7 = pa.array(data7, type=t7)
-
-    table = pa.Table.from_arrays([a1, a2, a3, a4, a5, a6, a7],
-                                 ['date32', 'date64', 'timestamp[us]',
-                                  'time32[s]', 'time64[us]',
-                                  'time32_from64[s]',
-                                  'timestamp[ns]'])
-
-    # date64 as date32
-    # time32[s] to time32[ms]
-    expected = pa.Table.from_arrays([a1, a1, a3, a4, a5, ex_a6, a7],
-                                    ['date32', 'date64', 'timestamp[us]',
-                                     'time32[s]', 'time64[us]',
-                                     'time32_from64[s]',
-                                     'timestamp[ns]'])
-
-    _check_roundtrip(table, expected=expected, version='2.0')
-
-    t0 = pa.timestamp('ms')
-    data0 = np.arange(4, dtype='int64')
-    a0 = pa.array(data0, type=t0)
-
-    t1 = pa.timestamp('us')
-    data1 = np.arange(4, dtype='int64')
-    a1 = pa.array(data1, type=t1)
-
-    t2 = pa.timestamp('ns')
-    data2 = np.arange(4, dtype='int64')
-    a2 = pa.array(data2, type=t2)
-
-    table = pa.Table.from_arrays([a0, a1, a2],
-                                 ['ts[ms]', 'ts[us]', 'ts[ns]'])
-    expected = pa.Table.from_arrays([a0, a1, a2],
-                                    ['ts[ms]', 'ts[us]', 'ts[ns]'])
-
-    # int64 for all timestamps supported by default
-    filename = tempdir / 'int64_timestamps.parquet'
-    _write_table(table, filename, version='2.0')
-    parquet_schema = pq.ParquetFile(filename).schema
-    for i in range(3):
-        assert parquet_schema.column(i).physical_type == 'INT64'
-    read_table = _read_table(filename)
-    assert read_table.equals(expected)
-
-    t0_ns = pa.timestamp('ns')
-    data0_ns = np.array(data0 * 1000000, dtype='int64')
-    a0_ns = pa.array(data0_ns, type=t0_ns)
-
-    t1_ns = pa.timestamp('ns')
-    data1_ns = np.array(data1 * 1000, dtype='int64')
-    a1_ns = pa.array(data1_ns, type=t1_ns)
-
-    expected = pa.Table.from_arrays([a0_ns, a1_ns, a2],
-                                    ['ts[ms]', 'ts[us]', 'ts[ns]'])
-
-    # int96 nanosecond timestamps produced upon request
-    filename = tempdir / 'explicit_int96_timestamps.parquet'
-    _write_table(table, filename, version='2.0',
-                 use_deprecated_int96_timestamps=True)
-    parquet_schema = pq.ParquetFile(filename).schema
-    for i in range(3):
-        assert parquet_schema.column(i).physical_type == 'INT96'
-    read_table = _read_table(filename)
-    assert read_table.equals(expected)
-
-    # int96 nanosecond timestamps implied by flavor 'spark'
-    filename = tempdir / 'spark_int96_timestamps.parquet'
-    _write_table(table, filename, version='2.0',
-                 flavor='spark')
-    parquet_schema = pq.ParquetFile(filename).schema
-    for i in range(3):
-        assert parquet_schema.column(i).physical_type == 'INT96'
-    read_table = _read_table(filename)
-    assert read_table.equals(expected)
-
-
-def test_timestamp_restore_timezone():
-    # ARROW-5888, restore timezone from serialized metadata
-    ty = pa.timestamp('ms', tz='America/New_York')
-    arr = pa.array([1, 2, 3], type=ty)
-    t = pa.table([arr], names=['f0'])
-    _check_roundtrip(t)
-
-
-@pytest.mark.pandas
-def test_list_of_datetime_time_roundtrip():
-    # ARROW-4135
-    times = pd.to_datetime(['09:00', '09:30', '10:00', '10:30', '11:00',
-                            '11:30', '12:00'])
-    df = pd.DataFrame({'time': [times.time]})
-    _roundtrip_pandas_dataframe(df, write_kwargs={})
-
-
-@pytest.mark.pandas
-def test_parquet_version_timestamp_differences():
-    i_s = pd.Timestamp('2010-01-01').value / 1000000000  # := 1262304000
-
-    d_s = np.arange(i_s, i_s + 10, 1, dtype='int64')
-    d_ms = d_s * 1000
-    d_us = d_ms * 1000
-    d_ns = d_us * 1000
-
-    a_s = pa.array(d_s, type=pa.timestamp('s'))
-    a_ms = pa.array(d_ms, type=pa.timestamp('ms'))
-    a_us = pa.array(d_us, type=pa.timestamp('us'))
-    a_ns = pa.array(d_ns, type=pa.timestamp('ns'))
-
-    names = ['ts:s', 'ts:ms', 'ts:us', 'ts:ns']
-    table = pa.Table.from_arrays([a_s, a_ms, a_us, a_ns], names)
-
-    # Using Parquet version 1.0, seconds should be coerced to milliseconds
-    # and nanoseconds should be coerced to microseconds by default
-    expected = pa.Table.from_arrays([a_ms, a_ms, a_us, a_us], names)
-    _check_roundtrip(table, expected)
-
-    # Using Parquet version 2.0, seconds should be coerced to milliseconds
-    # and nanoseconds should be retained by default
-    expected = pa.Table.from_arrays([a_ms, a_ms, a_us, a_ns], names)
-    _check_roundtrip(table, expected, version='2.0')
-
-    # Using Parquet version 1.0, coercing to milliseconds or microseconds
-    # is allowed
-    expected = pa.Table.from_arrays([a_ms, a_ms, a_ms, a_ms], names)
-    _check_roundtrip(table, expected, coerce_timestamps='ms')
-
-    # Using Parquet version 2.0, coercing to milliseconds or microseconds
-    # is allowed
-    expected = pa.Table.from_arrays([a_us, a_us, a_us, a_us], names)
-    _check_roundtrip(table, expected, version='2.0', coerce_timestamps='us')
-
-    # TODO: after pyarrow allows coerce_timestamps='ns', tests like the
-    # following should pass ...
-
-    # Using Parquet version 1.0, coercing to nanoseconds is not allowed
-    # expected = None
-    # with pytest.raises(NotImplementedError):
-    #     _roundtrip_table(table, coerce_timestamps='ns')
-
-    # Using Parquet version 2.0, coercing to nanoseconds is allowed
-    # expected = pa.Table.from_arrays([a_ns, a_ns, a_ns, a_ns], names)
-    # _check_roundtrip(table, expected, version='2.0', coerce_timestamps='ns')
-
-    # For either Parquet version, coercing to nanoseconds is allowed
-    # if Int96 storage is used
-    expected = pa.Table.from_arrays([a_ns, a_ns, a_ns, a_ns], names)
-    _check_roundtrip(table, expected,
-                     use_deprecated_int96_timestamps=True)
-    _check_roundtrip(table, expected, version='2.0',
-                     use_deprecated_int96_timestamps=True)
 
 
 def test_large_list_records():
@@ -1110,17 +530,6 @@ def test_sanitized_spark_field_names():
 
     expected_name = 'prohib______'
     assert result.schema[0].name == expected_name
-
-
-@pytest.mark.pandas
-def test_spark_flavor_preserves_pandas_metadata():
-    df = _test_dataframe(size=100)
-    df.index = np.arange(0, 10 * len(df), 10)
-    df.index.name = 'foo'
-
-    result = _roundtrip_pandas_dataframe(df, {'version': '2.0',
-                                              'flavor': 'spark'})
-    tm.assert_frame_equal(result, df)
 
 
 def test_fixed_size_binary():
@@ -1295,28 +704,6 @@ def test_empty_directory(tempdir, use_legacy_dataset):
     assert result.num_columns == 0
 
 
-@pytest.mark.pandas
-@parametrize_legacy_dataset
-def test_multiindex_duplicate_values(tempdir, use_legacy_dataset):
-    num_rows = 3
-    numbers = list(range(num_rows))
-    index = pd.MultiIndex.from_arrays(
-        [['foo', 'foo', 'bar'], numbers],
-        names=['foobar', 'some_numbers'],
-    )
-
-    df = pd.DataFrame({'numbers': numbers}, index=index)
-    table = pa.Table.from_pandas(df)
-
-    filename = tempdir / 'dup_multi_index_levels.parquet'
-
-    _write_table(table, filename)
-    result_table = _read_table(filename, use_legacy_dataset=use_legacy_dataset)
-    assert table.equals(result_table)
-
-    result_df = result_table.to_pandas()
-    tm.assert_frame_equal(result_df, df)
-
 
 @pytest.mark.pandas
 def test_write_error_deletes_incomplete_file(tempdir):
@@ -1341,31 +728,6 @@ def test_write_error_deletes_incomplete_file(tempdir):
         pass
 
     assert not filename.exists()
-
-
-@pytest.mark.pandas
-def test_noncoerced_nanoseconds_written_without_exception(tempdir):
-    # ARROW-1957: the Parquet version 2.0 writer preserves Arrow
-    # nanosecond timestamps by default
-    n = 9
-    df = pd.DataFrame({'x': range(n)},
-                      index=pd.date_range('2017-01-01', freq='1n', periods=n))
-    tb = pa.Table.from_pandas(df)
-
-    filename = tempdir / 'written.parquet'
-    try:
-        pq.write_table(tb, filename, version='2.0')
-    except Exception:
-        pass
-    assert filename.exists()
-
-    recovered_table = pq.read_table(filename)
-    assert tb.equals(recovered_table)
-
-    # Loss of data through coercion (without explicit override) still an error
-    filename = tempdir / 'not_written.parquet'
-    with pytest.raises(ValueError):
-        pq.write_table(tb, filename, coerce_timestamps='ms', version='2.0')
 
 
 @parametrize_legacy_dataset
@@ -1474,32 +836,6 @@ def test_list_of_binary_large_cell(use_legacy_dataset):
 
 @pytest.mark.pandas
 @parametrize_legacy_dataset
-def test_index_column_name_duplicate(tempdir, use_legacy_dataset):
-    data = {
-        'close': {
-            pd.Timestamp('2017-06-30 01:31:00'): 154.99958999999998,
-            pd.Timestamp('2017-06-30 01:32:00'): 154.99958999999998,
-        },
-        'time': {
-            pd.Timestamp('2017-06-30 01:31:00'): pd.Timestamp(
-                '2017-06-30 01:31:00'
-            ),
-            pd.Timestamp('2017-06-30 01:32:00'): pd.Timestamp(
-                '2017-06-30 01:32:00'
-            ),
-        }
-    }
-    path = str(tempdir / 'data.parquet')
-    dfx = pd.DataFrame(data).set_index('time', drop=False)
-    tdfx = pa.Table.from_pandas(dfx)
-    _write_table(tdfx, path)
-    arrow_table = _read_table(path, use_legacy_dataset=use_legacy_dataset)
-    result_df = arrow_table.to_pandas()
-    tm.assert_frame_equal(result_df, dfx)
-
-
-@pytest.mark.pandas
-@parametrize_legacy_dataset
 def test_parquet_nested_convenience(tempdir, use_legacy_dataset):
     # ARROW-1684
     df = pd.DataFrame({
@@ -1519,112 +855,6 @@ def test_parquet_nested_convenience(tempdir, use_legacy_dataset):
     read = pq.read_table(
         path, columns=['a', 'b'], use_legacy_dataset=use_legacy_dataset)
     tm.assert_frame_equal(read.to_pandas(), df)
-
-
-@pytest.mark.pandas
-@parametrize_legacy_dataset
-def test_backwards_compatible_index_naming(datadir, use_legacy_dataset):
-    expected_string = b"""\
-carat        cut  color  clarity  depth  table  price     x     y     z
- 0.23      Ideal      E      SI2   61.5   55.0    326  3.95  3.98  2.43
- 0.21    Premium      E      SI1   59.8   61.0    326  3.89  3.84  2.31
- 0.23       Good      E      VS1   56.9   65.0    327  4.05  4.07  2.31
- 0.29    Premium      I      VS2   62.4   58.0    334  4.20  4.23  2.63
- 0.31       Good      J      SI2   63.3   58.0    335  4.34  4.35  2.75
- 0.24  Very Good      J     VVS2   62.8   57.0    336  3.94  3.96  2.48
- 0.24  Very Good      I     VVS1   62.3   57.0    336  3.95  3.98  2.47
- 0.26  Very Good      H      SI1   61.9   55.0    337  4.07  4.11  2.53
- 0.22       Fair      E      VS2   65.1   61.0    337  3.87  3.78  2.49
- 0.23  Very Good      H      VS1   59.4   61.0    338  4.00  4.05  2.39"""
-    expected = pd.read_csv(io.BytesIO(expected_string), sep=r'\s{2,}',
-                           index_col=None, header=0, engine='python')
-    table = _read_table(
-        datadir / 'v0.7.1.parquet', use_legacy_dataset=use_legacy_dataset)
-    result = table.to_pandas()
-    tm.assert_frame_equal(result, expected)
-
-
-@pytest.mark.pandas
-@parametrize_legacy_dataset
-def test_backwards_compatible_index_multi_level_named(
-    datadir, use_legacy_dataset
-):
-    expected_string = b"""\
-carat        cut  color  clarity  depth  table  price     x     y     z
- 0.23      Ideal      E      SI2   61.5   55.0    326  3.95  3.98  2.43
- 0.21    Premium      E      SI1   59.8   61.0    326  3.89  3.84  2.31
- 0.23       Good      E      VS1   56.9   65.0    327  4.05  4.07  2.31
- 0.29    Premium      I      VS2   62.4   58.0    334  4.20  4.23  2.63
- 0.31       Good      J      SI2   63.3   58.0    335  4.34  4.35  2.75
- 0.24  Very Good      J     VVS2   62.8   57.0    336  3.94  3.96  2.48
- 0.24  Very Good      I     VVS1   62.3   57.0    336  3.95  3.98  2.47
- 0.26  Very Good      H      SI1   61.9   55.0    337  4.07  4.11  2.53
- 0.22       Fair      E      VS2   65.1   61.0    337  3.87  3.78  2.49
- 0.23  Very Good      H      VS1   59.4   61.0    338  4.00  4.05  2.39"""
-    expected = pd.read_csv(
-        io.BytesIO(expected_string), sep=r'\s{2,}',
-        index_col=['cut', 'color', 'clarity'],
-        header=0, engine='python'
-    ).sort_index()
-
-    table = _read_table(datadir / 'v0.7.1.all-named-index.parquet',
-                        use_legacy_dataset=use_legacy_dataset)
-    result = table.to_pandas()
-    tm.assert_frame_equal(result, expected)
-
-
-@pytest.mark.pandas
-@parametrize_legacy_dataset
-def test_backwards_compatible_index_multi_level_some_named(
-        datadir, use_legacy_dataset
-):
-    expected_string = b"""\
-carat        cut  color  clarity  depth  table  price     x     y     z
- 0.23      Ideal      E      SI2   61.5   55.0    326  3.95  3.98  2.43
- 0.21    Premium      E      SI1   59.8   61.0    326  3.89  3.84  2.31
- 0.23       Good      E      VS1   56.9   65.0    327  4.05  4.07  2.31
- 0.29    Premium      I      VS2   62.4   58.0    334  4.20  4.23  2.63
- 0.31       Good      J      SI2   63.3   58.0    335  4.34  4.35  2.75
- 0.24  Very Good      J     VVS2   62.8   57.0    336  3.94  3.96  2.48
- 0.24  Very Good      I     VVS1   62.3   57.0    336  3.95  3.98  2.47
- 0.26  Very Good      H      SI1   61.9   55.0    337  4.07  4.11  2.53
- 0.22       Fair      E      VS2   65.1   61.0    337  3.87  3.78  2.49
- 0.23  Very Good      H      VS1   59.4   61.0    338  4.00  4.05  2.39"""
-    expected = pd.read_csv(
-        io.BytesIO(expected_string),
-        sep=r'\s{2,}', index_col=['cut', 'color', 'clarity'],
-        header=0, engine='python'
-    ).sort_index()
-    expected.index = expected.index.set_names(['cut', None, 'clarity'])
-
-    table = _read_table(datadir / 'v0.7.1.some-named-index.parquet',
-                        use_legacy_dataset=use_legacy_dataset)
-    result = table.to_pandas()
-    tm.assert_frame_equal(result, expected)
-
-
-@pytest.mark.pandas
-@parametrize_legacy_dataset
-def test_backwards_compatible_column_metadata_handling(
-    datadir, use_legacy_dataset
-):
-    expected = pd.DataFrame(
-        {'a': [1, 2, 3], 'b': [.1, .2, .3],
-         'c': pd.date_range("2017-01-01", periods=3, tz='Europe/Brussels')})
-    expected.index = pd.MultiIndex.from_arrays(
-        [['a', 'b', 'c'],
-         pd.date_range("2017-01-01", periods=3, tz='Europe/Brussels')],
-        names=['index', None])
-
-    path = datadir / 'v0.7.1.column-metadata-handling.parquet'
-    table = _read_table(path, use_legacy_dataset=use_legacy_dataset)
-    result = table.to_pandas()
-    tm.assert_frame_equal(result, expected)
-
-    table = _read_table(
-        path, columns=['a'], use_legacy_dataset=use_legacy_dataset)
-    result = table.to_pandas()
-    tm.assert_frame_equal(result, expected[['a']].reset_index(drop=True))
 
 
 @pytest.mark.pandas
@@ -1746,34 +976,6 @@ def test_zlib_compression_bug(use_legacy_dataset):
     tm.assert_frame_equal(roundtrip.to_pandas(), table.to_pandas())
 
 
-@pytest.mark.pandas
-def test_merging_parquet_tables_with_different_pandas_metadata(tempdir):
-    # ARROW-3728: Merging Parquet Files - Pandas Meta in Schema Mismatch
-    schema = pa.schema([
-        pa.field('int', pa.int16()),
-        pa.field('float', pa.float32()),
-        pa.field('string', pa.string())
-    ])
-    df1 = pd.DataFrame({
-        'int': np.arange(3, dtype=np.uint8),
-        'float': np.arange(3, dtype=np.float32),
-        'string': ['ABBA', 'EDDA', 'ACDC']
-    })
-    df2 = pd.DataFrame({
-        'int': [4, 5],
-        'float': [1.1, None],
-        'string': [None, None]
-    })
-    table1 = pa.Table.from_pandas(df1, schema=schema, preserve_index=False)
-    table2 = pa.Table.from_pandas(df2, schema=schema, preserve_index=False)
-
-    assert not table1.schema.equals(table2.schema, check_metadata=True)
-    assert table1.schema.equals(table2.schema)
-
-    writer = pq.ParquetWriter(tempdir / 'merged.parquet', schema=schema)
-    writer.write_table(table1)
-    writer.write_table(table2)
-
 
 def test_empty_row_groups(tempdir):
     # ARROW-3020
@@ -1851,64 +1053,6 @@ def test_read_column_invalid_index():
             f.reader.read_column(index)
 
 
-@pytest.mark.pandas
-@parametrize_legacy_dataset
-def test_direct_read_dictionary(use_legacy_dataset):
-    # ARROW-3325
-    repeats = 10
-    nunique = 5
-
-    data = [
-        [util.rands(10) for i in range(nunique)] * repeats,
-
-    ]
-    table = pa.table(data, names=['f0'])
-
-    bio = pa.BufferOutputStream()
-    pq.write_table(table, bio)
-    contents = bio.getvalue()
-
-    result = pq.read_table(pa.BufferReader(contents),
-                           read_dictionary=['f0'],
-                           use_legacy_dataset=use_legacy_dataset)
-
-    # Compute dictionary-encoded subfield
-    expected = pa.table([table[0].dictionary_encode()], names=['f0'])
-    assert result.equals(expected)
-
-
-@pytest.mark.pandas
-@parametrize_legacy_dataset
-def test_direct_read_dictionary_subfield(use_legacy_dataset):
-    repeats = 10
-    nunique = 5
-
-    data = [
-        [[util.rands(10)] for i in range(nunique)] * repeats,
-    ]
-    table = pa.table(data, names=['f0'])
-
-    bio = pa.BufferOutputStream()
-    pq.write_table(table, bio)
-    contents = bio.getvalue()
-    result = pq.read_table(pa.BufferReader(contents),
-                           read_dictionary=['f0.list.item'],
-                           use_legacy_dataset=use_legacy_dataset)
-
-    arr = pa.array(data[0])
-    values_as_dict = arr.values.dictionary_encode()
-
-    inner_indices = values_as_dict.indices.cast('int32')
-    new_values = pa.DictionaryArray.from_arrays(inner_indices,
-                                                values_as_dict.dictionary)
-
-    offsets = pa.array(range(51), type='int32')
-    expected_arr = pa.ListArray.from_arrays(offsets, new_values)
-    expected = pa.table([expected_arr], names=['f0'])
-
-    assert result.equals(expected)
-    assert result[0].num_chunks == 1
-
 
 @parametrize_legacy_dataset
 def test_parquet_file_too_small(tempdir, use_legacy_dataset):
@@ -1927,154 +1071,6 @@ def test_parquet_file_too_small(tempdir, use_legacy_dataset):
         pq.read_table(path, use_legacy_dataset=use_legacy_dataset)
 
 
-@pytest.mark.pandas
-@parametrize_legacy_dataset
-def test_categorical_index_survives_roundtrip(use_legacy_dataset):
-    # ARROW-3652, addressed by ARROW-3246
-    df = pd.DataFrame([['a', 'b'], ['c', 'd']], columns=['c1', 'c2'])
-    df['c1'] = df['c1'].astype('category')
-    df = df.set_index(['c1'])
-
-    table = pa.Table.from_pandas(df)
-    bos = pa.BufferOutputStream()
-    pq.write_table(table, bos)
-    ref_df = pq.read_pandas(
-        bos.getvalue(), use_legacy_dataset=use_legacy_dataset).to_pandas()
-    assert isinstance(ref_df.index, pd.CategoricalIndex)
-    assert ref_df.index.equals(df.index)
-
-
-@pytest.mark.pandas
-@parametrize_legacy_dataset
-def test_categorical_order_survives_roundtrip(use_legacy_dataset):
-    # ARROW-6302
-    df = pd.DataFrame({"a": pd.Categorical(
-        ["a", "b", "c", "a"], categories=["b", "c", "d"], ordered=True)})
-
-    table = pa.Table.from_pandas(df)
-    bos = pa.BufferOutputStream()
-    pq.write_table(table, bos)
-
-    contents = bos.getvalue()
-    result = pq.read_pandas(
-        contents, use_legacy_dataset=use_legacy_dataset).to_pandas()
-
-    tm.assert_frame_equal(result, df)
-
-
-def _simple_table_write_read(table, use_legacy_dataset):
-    bio = pa.BufferOutputStream()
-    pq.write_table(table, bio)
-    contents = bio.getvalue()
-    return pq.read_table(
-        pa.BufferReader(contents), use_legacy_dataset=use_legacy_dataset
-    )
-
-
-@parametrize_legacy_dataset
-def test_dictionary_array_automatically_read(use_legacy_dataset):
-    # ARROW-3246
-
-    # Make a large dictionary, a little over 4MB of data
-    dict_length = 4000
-    dict_values = pa.array([('x' * 1000 + '_{}'.format(i))
-                            for i in range(dict_length)])
-
-    num_chunks = 10
-    chunk_size = 100
-    chunks = []
-    for i in range(num_chunks):
-        indices = np.random.randint(0, dict_length,
-                                    size=chunk_size).astype(np.int32)
-        chunks.append(pa.DictionaryArray.from_arrays(pa.array(indices),
-                                                     dict_values))
-
-    table = pa.table([pa.chunked_array(chunks)], names=['f0'])
-    result = _simple_table_write_read(table, use_legacy_dataset)
-
-    assert result.equals(table)
-
-    # The only key in the metadata was the Arrow schema key
-    assert result.schema.metadata is None
-
-
-def test_field_id_metadata():
-    # ARROW-7080
-    table = pa.table([pa.array([1], type='int32'),
-                      pa.array([[]], type=pa.list_(pa.int32())),
-                      pa.array([b'boo'], type='binary')],
-                     ['f0', 'f1', 'f2'])
-
-    bio = pa.BufferOutputStream()
-    pq.write_table(table, bio)
-    contents = bio.getvalue()
-
-    pf = pq.ParquetFile(pa.BufferReader(contents))
-    schema = pf.schema_arrow
-
-    # Expected Parquet schema for reference
-    #
-    # required group field_id=0 schema {
-    #   optional int32 field_id=1 f0;
-    #   optional group field_id=2 f1 (List) {
-    #     repeated group field_id=3 list {
-    #       optional int32 field_id=4 item;
-    #     }
-    #   }
-    #   optional binary field_id=5 f2;
-    # }
-
-    field_name = b'PARQUET:field_id'
-    assert schema[0].metadata[field_name] == b'1'
-
-    list_field = schema[1]
-    assert list_field.metadata[field_name] == b'2'
-
-    list_item_field = list_field.type.value_field
-    assert list_item_field.metadata[field_name] == b'4'
-
-    assert schema[2].metadata[field_name] == b'5'
-
-
-@pytest.mark.pandas
-@parametrize_legacy_dataset
-def test_pandas_categorical_na_type_row_groups(use_legacy_dataset):
-    # ARROW-5085
-    df = pd.DataFrame({"col": [None] * 100, "int": [1.0] * 100})
-    df_category = df.astype({"col": "category", "int": "category"})
-    table = pa.Table.from_pandas(df)
-    table_cat = pa.Table.from_pandas(df_category)
-    buf = pa.BufferOutputStream()
-
-    # it works
-    pq.write_table(table_cat, buf, version="2.0", chunk_size=10)
-    result = pq.read_table(
-        buf.getvalue(), use_legacy_dataset=use_legacy_dataset)
-
-    # Result is non-categorical
-    assert result[0].equals(table[0])
-    assert result[1].equals(table[1])
-
-
-@pytest.mark.pandas
-@parametrize_legacy_dataset
-def test_pandas_categorical_roundtrip(use_legacy_dataset):
-    # ARROW-5480, this was enabled by ARROW-3246
-
-    # Have one of the categories unobserved and include a null (-1)
-    codes = np.array([2, 0, 0, 2, 0, -1, 2], dtype='int32')
-    categories = ['foo', 'bar', 'baz']
-    df = pd.DataFrame({'x': pd.Categorical.from_codes(
-        codes, categories=categories)})
-
-    buf = pa.BufferOutputStream()
-    pq.write_table(pa.table(df), buf)
-
-    result = pq.read_table(
-        buf.getvalue(), use_legacy_dataset=use_legacy_dataset).to_pandas()
-    assert result.x.dtype == 'category'
-    assert (result.x.cat.categories == categories).all()
-    tm.assert_frame_equal(result, df)
 
 
 @parametrize_legacy_dataset
